% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils2.r
\name{macroInFocusGWConc}
\alias{macroInFocusGWConc}
\alias{macroInFocusGWConc.character}
\alias{macroInFocusGWConc.list}
\alias{macroInFocusGWConc.data.frame}
\title{INTERNAL/NON-OFFICIAL: Calculate the yearly and Xth percentile groundwater concentration from a MACROInFOCUS output.}
\usage{
macroInFocusGWConc(x, nbYrsWarmUp = 6L, yearsAvg = NULL, prob = 0.8,
  method = c("focus", "R")[1L], negToZero = TRUE, type = 7L,
  quiet = FALSE, ...)

\method{macroInFocusGWConc}{character}(x, nbYrsWarmUp = 6L,
  yearsAvg = NULL, prob = 0.8, method = c("focus", "R")[1L],
  negToZero = TRUE, type = 7L, quiet = FALSE, ...)

\method{macroInFocusGWConc}{list}(x, nbYrsWarmUp = 6L, yearsAvg = NULL,
  prob = 0.8, method = c("focus", "R")[1L], negToZero = TRUE,
  type = 7L, quiet = FALSE, ...)

\method{macroInFocusGWConc}{data.frame}(x, nbYrsWarmUp = 6L,
  yearsAvg = NULL, prob = 0.8, method = c("focus", "R")[1L],
  negToZero = TRUE, type = 7L, quiet = FALSE, ...)
}
\arguments{
\item{x}{Either a vector of character strings, a 
\code{\link[base]{data.frame}}, or a list of 
\code{\link[base]{data.frame}}s. If a vector of character 
strings, names (and possibly paths to) a \code{.bin} file 
output by MACROInFOCUS. The argument is passed internally 
to \code{\link[macroutils2:macroReadBin-methods]{macroReadBin}} (its 
\code{file}-argument). If a (list of) 
\code{\link[base]{data.frame}}(s), it should be imported from 
a \code{.bin} file output by MACROInFOCUS (for example 
with \code{\link[macroutils2:macroReadBin-methods]{macroReadBin}}).}

\item{nbYrsWarmUp}{Single integer values: Number of warm-up years that 
should be removed from the beginning of the model output.
A default of 6 years of warn-up are used in FOCUS.}

\item{yearsAvg}{Single integer values: Number of simulation years to 
"aggregate" when calculating yearly- or biennial- or 
triennial- (etc.) average concentrations. 
If \code{yearsAvg=1L}, the function calculates yearly-
average concentrations before calculating the Xth 
worst-case percentile. If \code{yearsAvg=2L}, the function 
calculates biennial-average concentrations. If 
\code{yearsAvg=3L}, the function calculates 
triennial-average concentrations (etc.). The default in 
FOCUS is to calculate yearly avegares when the pesticide 
is applied yearly, biennial-averages when the pesticide 
is applied every two years and triennial averages when 
the pesticide is applied every three years. When 
\code{yearsAvg} is \code{NULL} (the default), the function 
tries to automatically sets and control this parameter.}

\item{prob}{Single numeric value, between 0 and 1. Probability 
(percentile/100) of the worst case concentration 
that shall be calculated. In FOCUS, the yearly results 
are ordered by increasing yearly average concentrations 
before the percentile is calculated, as the average 
concentration of the two years closest to the Xth percentile 
(X always being 80, in FOCUS). Here, in practice, the 
index of the 1st and 2nd year used for calculating the 
average are selected as follow: 
\code{min_index = floor(prob *(number of sim years used))} and 
\code{max_index = ceiling(prob *(number of sim years used))},
but in cases \code{min_index} is indentical to \code{max_index}, 
then \code{max_index} is defined as \code{min_index + 1}, 
unless \code{prob} is 0 or 1 (to get the minimum 
or the maximum yearly concentrations, respectively). 
The number of simulatiob years used is equal to the total 
number of simulation years in \code{x} minus 
\code{nbYrsWarmUp}. In practice, what is calculated 
"a la FOCUS", when \code{prob = 0.8}, is an average 
between the 80th and the 85th percentile yearly 
concentrations. See FOCUS Groundwater main report p 475 
(reference above). Notice that the algorithm also calculates 
a Xth percentile concentration (\code{x = prob * 100}) 
using R function \code{\link[stats]{quantile}}, with 
its default parametrisation and quantile-calculation 
method (Note: see the help page of the function if you 
are interested to see how that percentile is obtained).}

\item{method}{Single character string. If \code{method = "focus"} (the default), 
the percentile is calculated with the default FOCUS method, 
that is the concentration derived from the cumulated 
yearly (or biennial or triennial) water and solute flow 
from the two years closest to the Xth percentile 
concentration (where \code{X = prob * 100}). If 
\code{method = "R"}, the concentration is calculated using 
\code{R} function \code{\link[stats]{quantile}}, calcuted 
directly on the yearly (or biennial or triennial) 
concentrations.}

\item{negToZero}{Single logical value. If \code{TRUE} (not the default) 
negative concentrations will be set to 0 (presumably like 
in MACROInFOCUS). If \code{FALSE}, they will not be set 
to 0, but if some of the concentrations used to calculate 
the Xth percentile (see \code{yearsXth}) are negative, 
a warning will be issued (so that the user knows that 
concentrations may differ from those in MACROInFOCUS).}

\item{type}{Single integer value. Only used when \code{method = "R"} 
(see above). See \code{\link[stats]{quantile}}.}

\item{quiet}{Single logical value. Set to \code{TRUE} to suppress the 
warning message. Default value to \code{FALSE}.}

\item{\dots}{Additional parameters passed to 
\code{\link[macroutils2:macroReadBin-methods]{macroReadBin}}, when \code{x} is 
a character string naming one or several files to be 
imported. Not used otherwise.}
}
\value{
Returns a \code{\link[base]{data.frame}} with one row per 
 input file (or table), and the following columns 
 \itemize{
   \item \code{concPercXth} (where \code{X} is 
     \code{prob * 100}): The Xth time-percentile of the 
     yearly, binennial or triennial (etc.) average pesticide 
     concentration in the water percolating 
     at the lower boundary of the soil. Derived from the 
     columns \code{TSOUT} and \code{TFLOWOUT}. Corresponds 
     to \code{CONC_PERC} in the table of yearly concentrations 
     (see below). For most scenario the target depth is 
     used rather than the percolation at the base of the 
     soil profile (see below).
   \item \code{concTLayerXth}  (where \code{X} is 
     \code{prob * 100}): The 
     Xth time-percentile of the yearly, binennial or triennial 
     average pesticide concentration in the water flowing 
     through layer T (T depending on the FOCUS-scenario, 
     also called "target depth" in MACROInFOCUS)  
     at depth Y. Derived from \code{SFLOW_}, 
     \code{SFLOWOUT_}, \code{WOUT_} and \code{WFLOWOUT_}, 
     after converting these variables to daily values and 
     cumulating them for each year. Corresponds 
     to \code{CONC_TLAYER} in the table of yearly concentrations 
     (see below).
   \item \code{tLayerAvgPerFrom} and \code{tLayerAvgPerTo}: 
     The 1st and 2nd periods corresponding to the Xth-percentile 
     concentration after ordering them by order of 
     increasing \code{CONC_TLAYER}. For example, in the 
     case of a yearly pesticide application, values of 
     10 and 15, respectively, indicate that the 10th and 
     the 15th simulation year, in chronological order and 
     after discarding the warm-up period, were used to 
     calculate the Xth time-percentile concentration at 
     the target depth.
   \item \code{percAvgPerFrom} and \code{percAvgPerTo}: 
     The 1st and 2nd periods corresponding to the Xth-percentile 
     concentration after ordering them by order of 
     increasing \code{CONC_PERC}. For example, in the 
     case of a yearly pesticide application, values of 
     10 and 15, respectively, indicate that the 10th and 
     the 15th simulation year, in chronological order and 
     after discarding the warm-up period, were used to 
     calculate the Xth time-percentile concentration at the 
     base of the soil.
   \item \code{avgIndexFrom} and \code{avgIndexTo}: 
     The 1st and 2nd period-indexes from which the 
     Xth-percentile concentration after ordering them by 
     order of increasing \code{CONC_TLAYER} or 
     \code{CONC_PERC}. For example, in the 
     case of a yearly pesticide application, values of 
     16 and 17, respectively, indicate that the 16th and 
     the 17th simulation year, after discarding the warm-up 
     period and ordering the periods in order of increasing 
     average concentration (at the target depth or at the 
     base of the soil profile), were used to 
     calculate the Xth time-percentile concentration at the 
     base of the soil and at the target depth.
   \item \code{percentile}: The percentile used to calculate 
     the concentrations. Equal to \code{prob * 100}. 
     See argument \code{prob} above.
   \item \code{method}: See argument \code{method} above.
   \item \code{nbSimYrsUsed}: Number of simulation years 
     used for the calculation, after discarding the warm-up 
     period.
   \item \code{nbYrsAvgPeriod}: Number of simulation years 
     aggregated to calculate the average concentration of 
     each "period". 1, 2 or 3 in cases of yearly, biennial 
     or triennial pesticide application freqency.
   \item \code{nbYrsWarmUp}: Number of simulation 
     years discared as a warm-up period.
   \item \code{fSolYLayerMacXth} (where \code{X} is 
     \code{prob * 100}): Experimental. The average 
     mass-fraction of the total solute flow due to transport 
     in the micropores, for the same two years as used 
     to calculate the Xth percentile of 
     \code{concTLayerXth}.
   \item \code{fSolTLayerMicXth} (where \code{X} is 
     \code{prob * 100}): Experimental. The average 
     mass-fraction of the total solute flow due to transport 
     in the macropores, for the same two years as used 
     to calculate the Xth percentile of 
     \code{concTLayerXth}.
   \item \code{file}: Path and name of the input file 
     from which the calculations were performed.
 }   
 In addition to this table, an \code{\link{attr}}ibute  
 is attached to the output table. This attribute is 
 names \code{more}, can be obtained by typing 
 \code{\link[base]{attr}( output, "more" )}, and contains 
 a table with all the yearly results for all the input files.
 It has the following columns: \code{"year"}, \code{"TSOUT"}, 
 \code{"TFLOWOUT"}, \code{"acc_SFLOW"}, \code{"acc_SFLOWOUT"}, 
 \code{"acc_WOUT"}, \code{"acc_WFLOWOUT"}, \code{"acc_WFLOWTOT"}, 
 \code{"acc_SFLOWTOT"}, \code{"CONC_PERC"}, \code{"CONC_TLAYER"}, 
 \code{"file"}. \code{"TSOUT"} and \code{"TFLOWOUT"} have 
 been de-accumulated and then re-accumulated for each simulation 
 \code{year}. All the columns starting with \code{acc_} 
 have been first converted to daily flow (instead of hourly 
 flow) and then accumulated for each simulation \code{year}. 
 \code{file} is the same as above. The total number of rows 
 should be \code{20 * length(x)} (20 simulation years times 
 number of input tables from MACROInFOCUS).
}
\description{
INTERNAL & NON-OFFICIAL: Calculate the yearly and Xth percentile 
 groundwater concentration from a MACROInFOCUS output. 
 \bold{WARNING} This function is \bold{not} part 
 of the official MACROInFOCUS program. It is provided 
 for test-purpose, without any guarantee or support from 
 the authors, CKB, SLU or KEMI. You are strongly recommended to 
 benchmark the function against a range of (official) 
 MACROInFOCUS simulation results, before you use the 
 function. You are also strongly recommended to inspect 
 the code of these functions before you use them. To 
 inspect the content of these functions, simply type 
 \code{body( macroutils2:::macroInFocusGWConc.data.frame )} 
 after you have loaded the package \code{macroutils2}.
}
\examples{

library( "macroutils2" ) 

#   Path to the file to be read
( filenm <- system.file( 
    "bintest/output_chat_winCer_GW-C_1kgHa_d298.bin", 
    package = "macroutils2", mustWork = TRUE ) )

res <- macroInFocusGWConc( x = filenm ) 

res 

attr( res, "more" ) 

#   Clean-up
rm( filenm, res )  


}
\references{
European Commission (2014) "Assessing Potential for 
 Movement of Active Substances and their Metabolites to 
 Ground Water in the EU Report of the FOCUS Ground Water 
 Work Group, EC Document Reference Sanco/13144/2010 
 version 3, 613 pp. \url{http://focus.jrc.ec.europa.eu/gw/docs/NewDocs/focusGWReportOct2014.pdf}
 See in particular the last sentence page 475.
}
\author{
Julien Moeys \email{jules_m78-soiltexture@yahooDOTfr}, 
 contributions from Stefan Reichenberger 
 \email{SReichenberger@knoellDOTcom}.
}
\keyword{internal}
