% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils2-package.R, R/macroutils2.r
\docType{package}
\name{macroutils2-package}
\alias{macroutils2-package}
\alias{macroutils}
\title{macroutils2: R Utility Functions for the MACRO and SOILNDB Models}
\description{
Read, write, convert, display and plot the content of binary files from the models MACRO and SOILNDB. Miscellaneous utility functions related MACRO, such as bug-fixes. macroutils2 is the continuation of macroutils, but there is not full backward compatibility (some functions' argument were changed or renamed). macroutils2 can read both MACRO output files and MACRO "intermediate files" for metabolites.

R utility functions for the MACRO (and SOIL) models. Reads, 
 writes, plot, view and converts MACRO binary files (input or 
 output data).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julienmoeys/macroutils2}
  \item Report bugs at \url{https://github.com/julienmoeys/macroutils2/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/julienmoeys/macroutils2}
  \item Report bugs at \url{https://github.com/julienmoeys/macroutils2/issues}
}

}
\author{
\strong{Maintainer}: Julien Moeys \email{jules_m78-soiltexture@@yahooDOTfr}

Authors:
\itemize{
  \item Kristian Persson
}

Other contributors:
\itemize{
  \item Centre for Chemical Pesticides (CKB), Swedish University of Agricultural Sciences (SLU) (Copyright 2010-2016) [copyright holder]
}


Julien MOEYS <jules_m78-soiltexture@yahooDOTfr>, 
 Kristian Persson <kristianDOTpersson@sluDOTse> 

 Maintainer: Julien MOEYS <jules_m78-soiltexture@yahooDOTfr>,
}
\keyword{internal}
\keyword{package}
