% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils2.r
\name{macroViewBin}
\alias{macroViewBin}
\title{Reads a MACRO/SOIL binary file and view it as a table.}
\usage{
macroViewBin(f, ...)
}
\arguments{
\item{f}{Single character strings or a single \code{\link{connection}} to
a binary file. If a vector character strings, it should be the name of the
binary file which the data are to be read from. The path of the file may be
provided as well, if \code{f} is not in the working directory.}

\item{\dots}{More arguments passed to \code{\link[macroutils2:macroReadBin-methods]{macroReadBin}}.}
}
\description{
Reads a MACRO/SOIL binary file and view it as a table.
}
