% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils2.r
\name{macroutilsFocusGWConc}
\alias{macroutilsFocusGWConc}
\alias{macroutilsFocusGWConc.character}
\alias{macroutilsFocusGWConc.list}
\alias{macroutilsFocusGWConc.data.frame}
\title{INTERNAL/NON-OFFICIAL: Calculate the yearly and Xth percentile groundwater concentration from a MACROInFOCUS output.}
\usage{
macroutilsFocusGWConc(x, nbYrsWarmUp = 6L, yearsAvg = NULL,
  prob = 0.8, method = c("focus", "R", "test")[1L], negToZero = TRUE,
  type = 7L, quiet = FALSE, massunits = 2L, ...)

\method{macroutilsFocusGWConc}{character}(x, nbYrsWarmUp = 6L,
  yearsAvg = NULL, prob = 0.8, method = c("focus", "R", "test")[1L],
  negToZero = TRUE, type = 7L, quiet = FALSE, massunits = 2L, ...)

\method{macroutilsFocusGWConc}{list}(x, nbYrsWarmUp = 6L,
  yearsAvg = NULL, prob = 0.8, method = c("focus", "R", "test")[1L],
  negToZero = TRUE, type = 7L, quiet = FALSE, massunits = 2L, ...)

\method{macroutilsFocusGWConc}{data.frame}(x, nbYrsWarmUp = 6L,
  yearsAvg = NULL, prob = 0.8, method = c("focus", "R", "test")[1L],
  negToZero = TRUE, type = 7L, quiet = FALSE, massunits = 2L, ...)
}
\arguments{
\item{x}{Either a vector of character strings, a 
\code{\link[base]{data.frame}}, or a list of 
\code{\link[base]{data.frame}}s. If a vector of character 
strings, names (and possibly paths to) a \code{.bin} file 
output by MACROInFOCUS. The argument is passed internally 
to \code{\link[macroutils2:macroReadBin-methods]{macroReadBin}} (its 
\code{file}-argument). If a (list of) 
\code{\link[base]{data.frame}}(s), it should be imported from 
a \code{.bin} file output by MACROInFOCUS (for example 
with \code{\link[macroutils2:macroReadBin-methods]{macroReadBin}}).}

\item{nbYrsWarmUp}{Single integer values: Number of warm-up years that 
should be removed from the beginning of the model output.
A default of 6 years of warn-up are used in FOCUS.}

\item{yearsAvg}{Single integer values: Number of simulation years to 
"aggregate" when calculating yearly- or biennial- or 
triennial- (etc.) average concentrations. 
If \code{yearsAvg=1L}, the function calculates yearly-
average concentrations before calculating the Xth 
worst-case percentile. If \code{yearsAvg=2L}, the function 
calculates biennial-average concentrations. If 
\code{yearsAvg=3L}, the function calculates 
triennial-average concentrations (etc.). The default in 
FOCUS is to calculate yearly avegares when the pesticide 
is applied yearly, biennial-averages when the pesticide 
is applied every two years and triennial averages when 
the pesticide is applied every three years. When 
\code{yearsAvg} is \code{NULL} (the default), the function 
tries to automatically sets and control this parameter.}

\item{prob}{Single numeric value, between 0 and 1. Probability 
(percentile/100) of the worst case concentration 
that shall be calculated. In FOCUS, the yearly results 
are ordered by increasing yearly average concentrations 
before the percentile is calculated, as the average 
concentration of the two years closest to the Xth percentile 
(X always being 80, in FOCUS). Here, in practice, the 
index of the 1st and 2nd year used for calculating the 
average are selected as follow: 
\code{min_index = floor(prob *(number of sim years used))} and 
\code{max_index = ceiling(prob *(number of sim years used))},
but in cases \code{min_index} is identical to \code{max_index}, 
then \code{max_index} is defined as \code{min_index + 1}, 
unless \code{prob} is 0 or 1 (to get the minimum 
or the maximum yearly concentrations, respectively). 
The number of simulation years used is equal to the total 
number of simulation years in \code{x} minus 
\code{nbYrsWarmUp}. In practice, what is calculated 
"a la FOCUS", when \code{prob = 0.8}, is an average 
between the 80th and the 85th percentile yearly 
concentrations. See FOCUS Groundwater main report p 475 
(reference above). Notice that the algorithm also calculates 
a Xth percentile concentration (\code{x = prob * 100}) 
using R function \code{\link[stats]{quantile}}, with 
its default parametrisation and quantile-calculation 
method (Note: see the help page of the function if you 
are interested to see how that percentile is obtained).}

\item{method}{Single character string. If \code{method = "focus"} (the default), 
the percentile is calculated with the default FOCUS method, 
that is the concentration derived from the cumulated 
yearly (or biennial or triennial) water and solute flow 
from the two years closest to the Xth percentile 
concentration (where \code{X = prob * 100}). If 
\code{method = "R"}, the concentration is calculated using 
\code{R} function \code{\link[stats]{quantile}}, calculated 
directly on the yearly (or biennial or triennial) 
concentrations. If \code{method = "test"}, it is expected 
that the simulation is a "short test" simulation, for example 
one year long, and a simple average concentration may be 
returned when a PEC-groundwater cannot be calculated. 
Only meant to be used when performing functional tests.}

\item{negToZero}{Single logical value. If \code{TRUE} (not the default) 
negative concentrations will be set to 0 (presumably like 
in MACROInFOCUS). If \code{FALSE}, they will not be set 
to 0, but if some of the concentrations used to calculate 
the Xth percentile (see \code{yearsXth}) are negative, 
a warning will be issued (so that the user knows that 
concentrations may differ from those in MACROInFOCUS).}

\item{type}{Single integer value. Only used when \code{method = "R"} 
(see above). See \code{\link[stats]{quantile}}.}

\item{quiet}{Single logical value. Set to \code{TRUE} to suppress the 
warning message. Default value to \code{FALSE}.}

\item{massunits}{Single integer value. Code for the mass unit of the simulation 
result in \code{x}. \code{1} is micro-grams, \code{2}is 
milligrams (the default in MACRO In FOCUS and thus in this 
function), \code{2} is grams and \code{4} is kilograms. 
Corresponds to the parameter \code{MASSUNITS} in MACRO.}

\item{\dots}{Additional parameters passed to 
\code{\link[macroutils2:macroReadBin-methods]{macroReadBin}}, when \code{x} is 
a character string naming one or several files to be 
imported. Not used otherwise.}
}
\value{
Returns a \code{\link[base]{list}} with the following items:
 \itemize{
   \item{"info\_general"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"conc\_percentile"}{The percentile used to 
         calculate the Predicted Environmental Concentration 
         (columns \code{ug\_per\_L} in items 
         \code{conc\_target\_layer} and \code{conc\_perc}, 
         below), in [\%].}
       \item{"rank\_period1"}{The rank of the first 
         simulation period used to calculate \code{ug\_per\_L},
         when ordered by increasing average concentration.}
       \item{"rank\_period2"}{The rank of the second 
         simulation period used to calculate \code{ug\_per\_L},
         when ordered by increasing average concentration.}
       \item{"method"}{See argument \code{method} above.}
       \item{"nb\_sim\_yrs\_used"}{Number of simulation years 
         used for the calculation, after discarding the warm-up 
         period.}
       \item{"nb\_yrs\_per\_period"}{Number of simulation years 
         aggregated to calculate the average concentration of 
         each "period". 1, 2 or 3 in cases of yearly, biennial 
         or triennial pesticide application frequency.}
       \item{"nb\_yrs\_warmup"}{Number of simulation 
         years discarded as a warm-up period.}
       \item{"neg\_conc\_set\_to\_0"}{See \code{negToZero} 
         above.}
     }  
   }  
   \item{"info\_period"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"period\_index"}{Index of the simulation 
         period when periods are sorted in chronological 
         order (i.e. \code{1} is the first or earliest 
         period).}
       \item{"from\_year"}{First year included in the period.}
       \item{"to\_year"}{Last year included in the period.}
     }  
   } 
   \item{"water\_target\_layer\_by\_period"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"period\_index"}{See above.}
       \item{"mm\_mic"}{Accumulated amount of water 
         passed through the micropores in target layer 
         over the period, downward (positive) or upward 
         (negative), in [mm] of water.}
       \item{"mm\_mac"}{Accumulated amount of water 
         passed through the macropores in target layer 
         over the period, downward (positive) or upward 
         (negative), in [mm] of water.}
       \item{"mm\_tot"}{Accumulated amount of water 
         passed through the target layer 
         over the period, downward (positive) or upward 
         (negative), in [mm] of water.}
     }  
   } 
   \item{"solute\_target\_layer\_by\_period"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"period\_index"}{See above.}
       \item{"mg\_per\_m2\_mic"}{Accumulated mass of 
         solute passed through the micropores, per square 
         meter, in target layer, over the period, 
         downward (positive) or upward (negative), in 
         [mg/ m2].}
       \item{"mg\_per\_m2\_mac"}{Accumulated mass of 
         solute passed through the macropores, per square 
         meter, in target layer, over the period, 
         downward (positive) or upward (negative), in 
         [mg/ m2].}
       \item{"mg\_per\_m2\_tot"}{Accumulated mass of 
         solute passed through the target layer, per square 
         meter, over the period, downward (positive) or 
         upward (negative), in [mg/ m2].}
       \item{"ug\_per\_L"}{Water-flow-weighted average 
         solute concentration over the period, in 
         [micro-grams/L] or [mg/m3]. In practice equal 
         to the accumulated solute mass divided by the 
         accumulated water flow, with appropriate unit 
         conversion.}
     }  
   } 
   \item{"water\_perc\_by\_period"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"period\_index"}{See above.}
       \item{"mm"}{Accumulated amount of water 
         passed through the bottom layer of the soil profile 
         over the period, downward (positive) or upward 
         (negative), in [mm] of water.}
     }  
   } 
   \item{"solute\_perc\_by\_period"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"period\_index"}{See above.}
       \item{"mg\_per\_m2"}{Accumulated mass of 
         solute passed through the bottom layer of the soil 
         profile , per square meter, over the period, 
         downward (positive) or upward (negative), in 
         [mg/ m2].}
       \item{"ug\_per\_L"}{Water-flow-weighted average 
         solute concentration over the period, in 
         [micro-grams/L] or [mg/m3]. In practice equal 
         to the accumulated solute mass divided by the 
         accumulated water flow, with appropriate unit 
         conversion.}
     }  
   } 
   \item{"conc\_target\_layer"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"ug\_per\_L"}{Xth percentile of the period-
         averaged solute concentrations in the target 
         layer, where X is equal to \code{conc\_percentile} 
         (See above).}
       \item{"ug\_per\_L\_rnd"}{Same as above, except that 
         the concentration is rounded to 2 digits after the 
         decimal mark, in scientific mode, in an attempt to 
         obtain the same value as MACRO In FOCUS graphical 
         interface.}
       \item{"index\_period1"}{Index of the first simulation 
         period used to calculate the Xth percentile of 
         the period-averaged solute concentrations. 
         Corresponds to the column \code{period\_index} in 
         the tables above.}
       \item{"index\_period2"}{Index of the second simulation 
         period used to calculate the Xth percentile of 
         the period-averaged solute concentrations. 
         Corresponds to the column \code{period\_index} in 
         the tables above.}
       \item{"f\_solute\_mac"}{Average fraction of solute 
         in the macropores corresponding to 
         \code{ug\_per\_L}, 0 meaning 0\% of the solute in 
         the macropres and 1 meaning 100\% of the solute 
         in the macropores.}
       \item{"f\_solute\_mic"}{Average fraction of solute 
         in the micropores corresponding to 
         \code{ug\_per\_L}, 0 meaning 0\% of the solute in 
         the micropres and 1 meaning 100\% of the solute 
         in the micropores.}
     }  
   } 
   \item{"conc\_perc"}{
     A \code{\link[base]{data.frame}} with the following columns:
     \itemize{
       \item{"ug\_per\_L"}{Xth percentile of the period-
         averaged solute concentrations percolated at the 
         bottom boundary of the soil profile, where X is 
         equal to \code{conc\_percentile} (See above).}
       \item{"ug\_per\_L\_rnd"}{Same as above, except that 
         the concentration is rounded to 2 digits after the 
         decimal mark, in scientific mode, in an attempt to 
         obtain the same value as MACRO In FOCUS graphical 
         interface.}
       \item{"index\_period1"}{Index of the first simulation 
         period used to calculate the Xth percentile of 
         the period-averaged solute concentrations. 
         Corresponds to the column \code{period\_index} in 
         the tables above.}
       \item{"index\_period2"}{Index of the second simulation 
         period used to calculate the Xth percentile of 
         the period-averaged solute concentrations. 
         Corresponds to the column \code{period\_index} in 
         the tables above.}
     }  
   } 
 }
}
\description{
INTERNAL & NON-OFFICIAL: Calculate the yearly and Xth percentile 
 groundwater concentration from a MACROInFOCUS output. 
 \bold{WARNING} This function is \bold{not} part 
 of the official MACROInFOCUS program. It is provided 
 for test-purpose, without any guarantee or support from 
 the authors, CKB, SLU or KEMI. You are strongly recommended to 
 benchmark the function against a range of (official) 
 MACROInFOCUS simulation results, before you use the 
 function. You are also strongly recommended to inspect 
 the code of these functions before you use them. To 
 inspect the content of these functions, simply type 
 \code{body( macroutils2:::macroutilsFocusGWConc.data.frame )} 
 after you have loaded the package \code{macroutils2}.
}
\examples{

library( "macroutils2" ) 

#   Path to the file to be read
( filenm <- system.file( 
    "bintest/output_chat_winCer_GW-C_1kgHa_d298.bin", 
    package = "macroutils2", mustWork = TRUE ) )

res <- macroutilsFocusGWConc( x = filenm ) 

res 

#   Clean-up
rm( filenm, res )  


}
\references{
European Commission (2014) "Assessing Potential for 
 Movement of Active Substances and their Metabolites to 
 Ground Water in the EU Report of the FOCUS Ground Water 
 Work Group, EC Document Reference Sanco/13144/2010 
 version 3, 613 pp. \url{http://focus.jrc.ec.europa.eu/gw/docs/NewDocs/focusGWReportOct2014.pdf}
 See in particular the last sentence page 475.
}
\author{
Julien Moeys \email{jules_m78-soiltexture@yahooDOTfr}, 
 contributions from Stefan Reichenberger 
 \email{SReichenberger@knoellDOTcom}.
}
